refactor: LottoTicketFactory.class 삭제 및 하위 클래스 수정
