--- src/main/java/lotto/domain/factory/LottoTicketFactory.java
@@ -1,8 +1,0 @@
-package lotto.domain.factory;
-
-import lotto.domain.LottoTicket;
-
-@FunctionalInterface
-public interface LottoTicketFactory {
-    LottoTicket create();
-}
