+++ src/test/java/lotto/domain/factory/AutoLottoTicketFactoryTest.java
@@ -1,7 +1,6 @@
-package lotto.domain;
+package lotto.domain.factory;
 
-import lotto.domain.factory.AutoLottoTicketFactory;
-import lotto.domain.factory.LottoTicketFactory;
+import lotto.domain.LottoTicket;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Java6Assertions.assertThat;
@@ -10,7 +9,7 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 class AutoLottoTicketFactoryTest {
     @Test
     void 예외없이_LottoTicket_생성되는지_확인() {
-        final LottoTicketFactory lottoTicketFactory = new AutoLottoTicketFactory((x) -> x);
+        final AutoLottoTicketFactory lottoTicketFactory = new AutoLottoTicketFactory((x) -> x);
 
         assertDoesNotThrow(lottoTicketFactory::create);
         assertThat(lottoTicketFactory.create()).isNotNull();
@@ -18,7 +17,7 @@ class AutoLottoTicketFactoryTest {
 
     @Test
     void shuffle_제대로_되는지_확인() {
-        final LottoTicketFactory lottoTicketFactory = new AutoLottoTicketFactory((x) -> x);
+        final AutoLottoTicketFactory lottoTicketFactory = new AutoLottoTicketFactory((x) -> x);
         final LottoTicket lottoTicket1 = lottoTicketFactory.create();
         final LottoTicket lottoTicket2 = lottoTicketFactory.create();
